Create base class for simple GLSL shows.
